feat: add scoring, team, and tournament routers with comprehensive CRUD operations
- Implemented scoring router with procedures for match scoring setup, starting innings, saving lineups, and recording deliveries.
- Created team router for managing teams, including registration, unassignment, and updates.
- Developed tournament router for handling tournaments, stages, teams, and organizations with appropriate admin access controls.
- Added input validation using Zod schemas for various operations across routers.
- Enhanced error handling for CRUD operations to provide meaningful feedback.
- Updated score-a-ball component to remove unused imports.
