+++ apps/server/src/routers/index.test.ts
@@ -1,0 +1,115 @@
+import { describe, expect, it } from "bun:test";
+import { appRouter } from "@/routers";
+
+const EXPECTED_PROCEDURE_NAMES = [
+  "healthCheck",
+  "privateData",
+  "currentUserRole",
+  "liveMatches",
+  "completedMatches",
+  "createMatch",
+  "getMatchById",
+  "organizations",
+  "matchFormats",
+  "liveTournaments",
+  "tournaments",
+  "managementTournaments",
+  "managementTournamentById",
+  "createTournament",
+  "createTournamentFromScratch",
+  "updateTournamentFromScratch",
+  "updateTournament",
+  "deleteTournament",
+  "createOrganization",
+  "createMatchFormat",
+  "tournamentTeams",
+  "tournamentTeamById",
+  "createTournamentTeam",
+  "updateTournamentTeam",
+  "deleteTournamentTeam",
+  "tournamentStages",
+  "tournamentStageById",
+  "tournamentStageGroups",
+  "tournamentStageGroupById",
+  "tournamentStageTeamEntries",
+  "tournamentStageTeamEntryById",
+  "createTournamentStageTeamEntry",
+  "updateTournamentStageTeamEntry",
+  "deleteTournamentStageTeamEntry",
+  "tournamentStageAdvancements",
+  "tournamentStageAdvancementById",
+  "createTournamentStageAdvancement",
+  "updateTournamentStageAdvancement",
+  "deleteTournamentStageAdvancement",
+  "tournamentStructure",
+  "createTournamentStage",
+  "updateTournamentStage",
+  "deleteTournamentStage",
+  "createTournamentStageGroup",
+  "updateTournamentStageGroup",
+  "deleteTournamentStageGroup",
+  "stageGroupsByStage",
+  "tournamentView",
+  "tournamentFixtures",
+  "tournamentStandings",
+  "setStagePointsConfig",
+  "createDraftFixtureMatch",
+  "updateDraftFixtureMatch",
+  "deleteDraftFixtureMatch",
+  "autoGenerateFixtures",
+  "autoGenerateNextSwissRound",
+  "publishFixtureMatches",
+  "seedTournamentTemplate",
+  "createFixtureDraft",
+  "createFixtureRound",
+  "publishFixtureVersion",
+  "validateFixtureConflicts",
+  "players",
+  "playersWithCurrentTeams",
+  "onboardingStatus",
+  "markOnboardingSeen",
+  "createOwnPlayerProfile",
+  "claimablePlayers",
+  "sendClaimOtp",
+  "verifyClaimOtpAndLink",
+  "createPlayer",
+  "bulkImportPlayers",
+  "updatePlayer",
+  "deletePlayer",
+  "teams",
+  "searchTeamsByName",
+  "getTeamById",
+  "teamTournaments",
+  "getTournamentTeamRoster",
+  "listTeamStats",
+  "getTeamStatsById",
+  "getTeamTournamentStats",
+  "createTeam",
+  "registerTournamentTeamPlayer",
+  "unassignTournamentTeamPlayer",
+  "reassignTournamentTeamPlayer",
+  "updateTeam",
+  "deleteTeam",
+  "getMatchScoringSetup",
+  "startMatchScoring",
+  "saveMatchLineup",
+  "initializeMatchScoring",
+  "startScoringInnings",
+  "recordScoringDelivery",
+  "updateScoringDelivery",
+  "deleteScoringDelivery",
+  "closeCurrentScoringInnings",
+  "saveScoringDelivery",
+  "createNextScoringDelivery",
+  "endScoringInnings",
+  "getMatchScorecard",
+  "getBallsOfSameOver",
+] as const;
+
+describe("appRouter", () => {
+  it("preserves the flat rpc contract surface", () => {
+    expect(Object.keys(appRouter).sort()).toEqual(
+      [...EXPECTED_PROCEDURE_NAMES].sort()
+    );
+  });
+});
